refactor: use the same code to emit cot and react messages
